+++ CHANGES.md
@@ -7,11 +7,14 @@ Fast and flexible PHP library for text tables.
 ## Changes
 
 * @dev
-  * Added `MsDosRenderer` producing tables reminiscent of ancient MS DOS times. 
+  * Added `MsDosRenderer` producing tables reminiscent of ancient MS DOS times.
+  * Added `renderAsString()` to `RendererContract`
+
 
 * 2.1.0 (2022-10-21)
   * Added `FancyRenderer` producing nicer looking table frames. It's now default one.
   * Former `DefaultRenderer` is now `PlusMinusRenderer`.
+
 
 * 2.0.0 (2022-10-20)
   * Main class name changed to `TextTable`.